Fixed issues and updated functionality.
Created view to list all authors in the Pubs database for clicking the button to populate ListBox with authors.
Used an ALTER view query statement to update the hard-coded author ID in the WHERE clause of the vAuthorTitles view.
Listing the titles by the selected author now directly uses the vAuthorTitles view after ALTER statement to change author ID to selected author ID in ListBox.
Minor spelling fixes in comments.